requirements.txt
Django==3.0.6
psycopg2-binary>=2.8
mysqlclient
markdown
# beautifulsoup4==4.8.1
django-cors-headers==3.2.0
django-extensions==2.2.5
django-filter==2.2.0
django-webpack-loader==0.6.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
django-allauth
gunicorn==20.0.4
Pillow==6.2.1
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
